requirements/development.txt
-r base.txt

flake8>=3.6.0
ipython>=7.1.1
isort>=4.3.4
pywatchman>=1.4.1

django-debug-toolbar>=1.10.1